requirements.txt
numpy
scipy
matplotlib
pandas
numba
torch
torchinfo
scikit-build
snakeviz
memory_profiler
control
pyyaml